+++ torch_geometric/distributed/rpc.py
@@ -1,19 +1,17 @@
-import atexit
 import logging
 import threading
 from abc import ABC, abstractmethod
-from typing import Callable, Dict, List, Optional
+from typing import Any, Callable, Dict, List, Optional
 
+from torch._C._distributed_rpc import _is_current_rpc_agent_set
 from torch.distributed import rpc
 
-import torch_geometric.typing
 from torch_geometric.distributed.dist_context import DistContext, DistRole
 
 _rpc_init_lock = threading.RLock()
 
 
 def rpc_is_initialized() -> bool:
-    from torch._C._distributed_rpc import _is_current_rpc_agent_set
     return _is_current_rpc_agent_set()
 
 
@@ -24,7 +22,7 @@ def rpc_require_initialized(func: Callable) -> Callable:
 
 
 @rpc_require_initialized
-def global_all_gather(obj, timeout: Optional[int] = None):
+def global_all_gather(obj, timeout: Optional[int] = None) -> Any:
     r"""Gathers objects from all groups in a list."""
     if timeout is None:
         return rpc.api._all_gather(obj)
@@ -32,21 +30,21 @@ def global_all_gather(obj, timeout: Optional[int] = None):
 
 
 @rpc_require_initialized
-def global_barrier(timeout: Optional[int] = None):
+def global_barrier(timeout: Optional[int] = None) -> None:
     r"""Block until all local and remote RPC processes."""
     try:
         global_all_gather(obj=None, timeout=timeout)
     except RuntimeError:
-        logging.error("Failed to respond to global barrier")
+        logging.error('Failed to respond to global barrier')
 
 
 def init_rpc(
     current_ctx: DistContext,
-    rpc_worker_names: Dict[DistRole, List[str]],
     master_addr: str,
     master_port: int,
     num_rpc_threads: int = 16,
-    rpc_timeout: int = 240,
+    rpc_timeout: float = 240.0,
+    rpc_worker_names: Optional[Dict[DistRole, List[str]]] = None,
 ):
     with _rpc_init_lock:
         if rpc_is_initialized():
@@ -70,33 +68,18 @@ def init_rpc(
             rpc_backend_options=options,
         )
 
-        gathered_results = global_all_gather(
-            obj=(current_ctx.role, current_ctx.world_size, current_ctx.rank),
-            timeout=rpc_timeout,
-        )
-
-        for worker_name, (role, world_size, rank) in gathered_results.items():
-            worker_list = rpc_worker_names.get(role, None)
-            if worker_list is None:
-                worker_list = [None for _ in range(world_size)]
-            else:
-                if len(worker_list) != world_size:
-                    raise RuntimeError(f"Inconsistent world size found in "
-                                       f"'init_rpc' (got {len(worker_list)})")
-
-            worker_list[rank] = worker_name
-            rpc_worker_names[role] = worker_list
-
         global_barrier(timeout=rpc_timeout)
 
 
-def shutdown_rpc(graceful: bool = True):
-    if rpc_is_initialized():
-        rpc.shutdown(graceful)
-
-
-if not torch_geometric.typing.WITH_WINDOWS:
-    atexit.register(shutdown_rpc, False)
+def shutdown_rpc(id: str = None, graceful: bool = True,
+                 timeout: float = 240.0):
+    with _rpc_init_lock:
+        if rpc_is_initialized():
+            global_barrier(timeout=timeout)
+            rpc.shutdown(graceful, timeout)
+            logging.debug(f'Closed RPC in {id} (graceful={graceful})')
+        else:
+            logging.error(f'RPC in {id} not initialized.')
 
 
 class RPCRouter:
@@ -104,7 +87,7 @@ class RPCRouter:
     def __init__(self, partition_to_workers: List[List[str]]):
         for pid, rpc_worker_list in enumerate(partition_to_workers):
             if len(rpc_worker_list) == 0:
-                raise ValueError("No RPC worker is in worker list")
+                raise ValueError('No RPC worker is in worker list')
         self.partition_to_workers = partition_to_workers
         self.rpc_worker_indices = [0 for _ in range(len(partition_to_workers))]